# pos_app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: POS Multi-Sucursal API
  version: 1.0.0
  description: Punto de venta y control de inventario multi-sucursal.
paths:
...
  /api/v1/auth/me/:
    get:
      operationId: auth_me
      description: Devuelve el usuario autenticado.
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
...
  /api/v1/branches/:
    get:
      operationId: branches_list
      description: Sucursales del negocio, ordenadas por nombre.
      parameters:
      - in: query
        name: active
        schema:
          type: boolean
        description: Solo sucursales activas.
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - branches
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBranchList'
          description: ''
    post:
      operationId: branches_create
      description: Crea una sucursal con stock cero de todos los productos.
      tags:
      - branches
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BranchCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BranchCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
          description: ''
  /api/v1/branches/{code}/:
    get:
      operationId: branches_retrieve
      description: Lectura para SUPERVISOR y MANAGER; escritura solo para MANAGER.
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - branches
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
          description: ''
    patch:
      operationId: branches_partial_update
      description: Cambia el nombre o activa/desactiva la sucursal. El código no cambia.
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - branches
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedBranchUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedBranchUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedBranchUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
          description: ''
...
  /api/v1/users/:
    get:
      operationId: users_list
      description: Usuarios del sistema, ordenados por nombre de usuario.
      parameters:
      - in: query
        name: active
        schema:
          type: boolean
        description: Solo usuarios activos.
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
          description: ''
    post:
      operationId: users_create
      description: Crea un usuario con su contraseña inicial.
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /api/v1/users/{id}/:
    get:
      operationId: users_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - users
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    patch:
      operationId: users_partial_update
      description: Actualiza los campos enviados; con `password` asigna una contraseña
        nueva.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
components:
  schemas:
...
    Branch:
      type: object
      properties:
        code:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
        active:
          type: boolean
          readOnly: true
      required:
      - active
      - code
      - name
    BranchCreate:
      type: object
      properties:
        code:
          type: string
          description: Identificador estable, p. ej. `CENTRO`. No se puede cambiar.
          maxLength: 20
        name:
          type: string
          maxLength: 100
      required:
      - code
      - name
...
    PaginatedBranchList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Branch'
...
    PaginatedUserList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/User'
    PatchedBranchUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        active:
          type: boolean
...
    PatchedUserUpdate:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 150
        role:
          $ref: '#/components/schemas/RoleEnum'
        assigned_branch:
          type: string
          nullable: true
          description: Código de la sucursal; null = todas (solo MANAGER).
          maxLength: 20
        is_active:
          type: boolean
        password:
          type: string
          writeOnly: true
          minLength: 8
...
    RoleEnum:
      enum:
      - MANAGER
      - SUPERVISOR
      type: string
      description: |-
        * `MANAGER` - Gerente
        * `SUPERVISOR` - Supervisor
...
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        assigned_branch:
          type: string
          readOnly: true
          nullable: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - assigned_branch
      - full_name
      - id
      - is_active
      - role
      - username
    UserCreate:
      type: object
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 8
        full_name:
          type: string
          maxLength: 150
        role:
          $ref: '#/components/schemas/RoleEnum'
        assigned_branch:
          type: string
          nullable: true
          description: Código de la sucursal; null = todas (solo MANAGER).
          maxLength: 20
      required:
      - assigned_branch
      - full_name
      - password
      - role
      - username
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT